Add coveralls support
Using Slather to generate coverage data for Coveralls, this adds all the
relevant files to support this. Mainly: .coveralls.yml, .slather.yml and
a new post test step in the circle.yml file. The COVERAGE_ACCESS_TOKEN
is added as an environment variable here:
https://circleci.com/gh/willowtreeapps/PinkyPromise/edit#env-vars

Coverage statistcs can be seen here: https://coveralls.io/github/willowtreeapps/PinkyPromise

There appears to be an issue with either Slather or Coveralls in sending
the correct Circle build number. Likely will address that in a later
bugfix branch.